refactored static header into shared library
